// src/app/core/Models/leads/ilead-filter.ts
export interface ILeadCounteries {
  clientLocation: string;
  count: number;
}

export interface ILeadCountry {
  clientLocation: string;
  count: number;
}

export interface ILeadCity {
  clientCity: string;
  count: number;
}

export interface ISourceCompany {
  sourceCompany: string;
  count: number;
}

export interface IChannel {
  entryChannel: string;
  count: number;
}

export interface ICampSource {
  clientWebsite: string;
  count: number;
}